backend/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
dashscope>=1.14.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
redis>=5.0.0
